+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Improvements
 
+- Support upcoming Writing Tools with Apple Intelligence on macOS Sequoia.
 - [beta] Adjust the window's appearance.
 - [beta] Add the Czech, Dutch, and Portuguese localizations for strings newly added in CotEditor 5.0.0.